+++ scratch.js
@@ -1,5 +1,5 @@
 var nlp = require('./src/index');
-nlp.verbose('tagger');
+// nlp.verbose('tagger');
 
 // let doc = nlp('100+').debug();
 
@@ -15,14 +15,12 @@ nlp.verbose('tagger');
 // nlp('Jim is nice, funny, cool').match('is #Adjective+').out();
 
 //isQuestion(), .not(Text)
-let doc = nlp(`Google's bite`);
-doc = doc.normalize({
-  possessives: true
-});
-// doc.debug();
 
-// console.log(doc.sentences().isQuestion().out('array'));
-// console.log(doc.sentences()[0].mainVerb());
-// let questions = doc.questions();
-// let sentences = doc.not(questions);
-// console.log(sentences.out('array'));
+let text = `
+It's free for 4 and free for 5
+`;
+let doc = nlp(text);
+let arr = doc.ngrams({
+  max: 3
+}).data();
+console.log(arr.find(o => o.normal === 'free for'));