+++ changelog.md
@@ -46,6 +46,9 @@ compromise uses semver, and pushes to npm frequently
 - slightly different behavior for wildcards in capture-groups [pull/472](https://github.com/spencermountain/compromise/pull/472)
 - `.possessives()` subset + `#Possessive` tagging fixes
 - hide massive `world` output for console.log of a term
+##### 11.8.0
+- add `nouns().toPossessive()`
+- ngrams now remove empty-terms in contractions - fixes counting issue [#476](https://github.com/spencermountain/compromise/issues/476)
 
 ##### 11.0.0
 - support for persistent lexicon/tagset changes